refactor(workspace): 拆出工作区子标签/分屏网格/侧边栏与看板组件
- 新增 components/workspace/：WorkspaceSubTabBar、WorkspaceTerminalGrid、WorkspaceSessionContent、WorkspaceSidePanes、WorkspaceDashboardSection、WorkspaceDiffModals、workspaceTypes
- SessionWorkspace.tsx 1433 -> 336 行
